Remove `ServiceTestConstants` and refactor test imports to use `TestConstants`.
