+++ src/test/java/com/chertiavdev/bookingapp/data/builders/AccommodationTestDataBuilder.java
@@ -2,19 +2,19 @@ package com.chertiavdev.bookingapp.data.builders;
 
 import static com.chertiavdev.bookingapp.model.Accommodation.Type.APARTMENT;
 import static com.chertiavdev.bookingapp.model.Accommodation.Type.HOUSE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_AVAILABILITY;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_DAILY_RATE_10050;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_DAILY_RATE_7550;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_DEFAULT_AMENITIES;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_SIZE_STUDIO;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_APARTMENT_NUMBER_25;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_APARTMENT_NUMBER_26;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_CITY_KYIV;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_HOUSE_NUMBER_15B;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_STREET_KHRESHCHATYK;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.AVAILABILITY_THRESHOLD;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_2;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_AVAILABILITY;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_DAILY_RATE_10050;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_DAILY_RATE_7550;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_DEFAULT_AMENITIES;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_SIZE_STUDIO;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_APARTMENT_NUMBER_25;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_APARTMENT_NUMBER_26;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_CITY_KYIV;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_HOUSE_NUMBER_15B;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_STREET_KHRESHCHATYK;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.AVAILABILITY_THRESHOLD;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_2;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createPage;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createPageResponse;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestAccommodation;
