+++ src/test/java/com/chertiavdev/bookingapp/data/builders/AmenityTestDataBuilder.java
@@ -1,12 +1,12 @@
 package com.chertiavdev.bookingapp.data.builders;
 
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.AMENITY_AIR_CONDITIONING_HEATING;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.AMENITY_FREE_WIFI;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.AMENITY_TELEVISION;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.AMENITY_UPDATED_NAME;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_2;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_3;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.AMENITY_AIR_CONDITIONING_HEATING;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.AMENITY_FREE_WIFI;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.AMENITY_TELEVISION;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.AMENITY_UPDATED_NAME;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_2;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_3;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestAmenity;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestAmenityRequestDto;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.mapAmenityToDto;
