+++ src/test/java/com/chertiavdev/bookingapp/data/builders/BookingTestDataBuilder.java
@@ -3,11 +3,11 @@ package com.chertiavdev.bookingapp.data.builders;
 import static com.chertiavdev.bookingapp.model.Booking.Status.CANCELED;
 import static com.chertiavdev.bookingapp.model.Booking.Status.CONFIRMED;
 import static com.chertiavdev.bookingapp.model.Booking.Status.PENDING;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_DAYS_UNTIL_CHECKOUT;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_2;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_4;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_5;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_DAYS_UNTIL_CHECKOUT;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_2;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_4;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_5;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createPage;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createPageResponse;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestBooking;
