+++ src/test/java/com/chertiavdev/bookingapp/data/builders/UserTestDataBuilder.java
@@ -1,19 +1,19 @@
 package com.chertiavdev.bookingapp.data.builders;
 
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.INVALID_USER_AUTH_PASSWORD;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_2;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_3;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USERNAME_FIRST_JOHN;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USERNAME_FIRST_SANSA;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USERNAME_LAST_DOE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USERNAME_LAST_STARK;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USER_EMAIL_JOHN;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USER_EMAIL_SANSA;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USER_FIRST_USERNAME_UPDATED;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USER_LAST_USERNAME_UPDATED;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.USER_PASSWORD;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.VALID_USER_AUTH_PASSWORD;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.INVALID_USER_AUTH_PASSWORD;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_2;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_3;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USERNAME_FIRST_JOHN;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USERNAME_FIRST_SANSA;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USERNAME_LAST_DOE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USERNAME_LAST_STARK;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USER_EMAIL_JOHN;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USER_EMAIL_SANSA;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USER_FIRST_USERNAME_UPDATED;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USER_LAST_USERNAME_UPDATED;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.USER_PASSWORD;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.VALID_USER_AUTH_PASSWORD;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestUser;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestUserLoginRequestDto;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createTestUserRegisterRequest;
