+++ src/test/java/com/chertiavdev/bookingapp/service/impl/BookingServiceImplTest.java
@@ -4,31 +4,31 @@ import static com.chertiavdev.bookingapp.model.Role.RoleName.ADMIN;
 import static com.chertiavdev.bookingapp.util.helpers.NotificationUtils.bookingNotificationForAdmins;
 import static com.chertiavdev.bookingapp.util.helpers.NotificationUtils.bookingNotificationToUser;
 import static com.chertiavdev.bookingapp.util.helpers.NotificationUtils.buildBookingExpiredAlert;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_APARTMENT_NUMBER_25;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_CITY_KYIV;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_HOUSE_NUMBER_15B;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ADDRESS_STREET_KHRESHCHATYK;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_ACTION_CANCELED;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_ACTION_CREATED;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_ALREADY_CANCELLED_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_HAS_PENDING_PAYMENTS_COUNT;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_NOT_FOUND_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_NO_EXPIRED_BOOKINGS_TODAY;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_PENDING_PAYMENTS_COUNT;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_PENDING_PAYMENT_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_UNAVAILABLE_ACCOMMODATION_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_UPDATE_ERROR_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_UPDATE_STATUS_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_USER_HAS_PAYMENT_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.INVALID_TEST_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SEARCH_STATUS_KEY;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SEARCH_USER_ID_KEY;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_NOT_BE_NULL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_APARTMENT_NUMBER_25;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_CITY_KYIV;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_HOUSE_NUMBER_15B;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ADDRESS_STREET_KHRESHCHATYK;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_ACTION_CANCELED;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_ACTION_CREATED;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_ALREADY_CANCELLED_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_HAS_PENDING_PAYMENTS_COUNT;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_NOT_FOUND_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_NO_EXPIRED_BOOKINGS_TODAY;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_PENDING_PAYMENTS_COUNT;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_PENDING_PAYMENT_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_UNAVAILABLE_ACCOMMODATION_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_UPDATE_ERROR_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_UPDATE_STATUS_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_USER_HAS_PAYMENT_MESSAGE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.CONTENT_OF_THE_PAGE_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.EXCEPTION_MESSAGE_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.INVALID_TEST_ID;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAGE_SIZE_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SEARCH_STATUS_KEY;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SEARCH_USER_ID_KEY;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.TOTAL_ELEMENTS_IN_THE_PAGE_DO_NOT_MATCH_THE_EXPECTED_VALUE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.TOTAL_NUMBER_OF_PAGES_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
 import static com.chertiavdev.bookingapp.utils.helpers.ServiceTestUtils.createAddressString;
