+++ src/test/java/com/chertiavdev/bookingapp/service/impl/PaymentServiceImplTest.java
@@ -4,23 +4,23 @@ import static com.chertiavdev.bookingapp.model.Role.RoleName.ADMIN;
 import static com.chertiavdev.bookingapp.util.constants.TelegramNotificationConstants.PAYMENT_CANCELLED_NOTIFICATION;
 import static com.chertiavdev.bookingapp.util.constants.TelegramNotificationConstants.PAYMENT_NOTIFICATION;
 import static com.chertiavdev.bookingapp.util.constants.TelegramNotificationConstants.PAYMENT_NOT_COMPLETED_NOTIFICATION;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.CAN_T_RETRIEVE_SESSION_BY_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.EXPECTED_BOOKING_STATUS_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.EXPECTED_PAYMENT_STATUS_PAID_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.INVALID_TEST_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_RENEWAL_INVALID_STATUS_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_RENEWAL_INVALID_USER_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_SESSION_NOT_FOUND_BY_ID_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_SESSION_NOT_FOUND_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_SESSION_PENDING_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_2;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_ID_RETRIEVAL_ERROR_TEMPLATE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_NOT_BE_NULL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.CAN_T_RETRIEVE_SESSION_BY_ID;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.CONTENT_OF_THE_PAGE_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.EXCEPTION_MESSAGE_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.EXPECTED_BOOKING_STATUS_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.EXPECTED_PAYMENT_STATUS_PAID_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.INVALID_TEST_ID;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAGE_SIZE_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_RENEWAL_INVALID_STATUS_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_RENEWAL_INVALID_USER_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_SESSION_NOT_FOUND_BY_ID_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_SESSION_NOT_FOUND_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_SESSION_PENDING_ID;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_2;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_ID_RETRIEVAL_ERROR_TEMPLATE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.TOTAL_ELEMENTS_IN_THE_PAGE_DO_NOT_MATCH_THE_EXPECTED_VALUE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.TOTAL_NUMBER_OF_PAGES_DOES_NOT_MATCH_THE_EXPECTED_VALUE;
 import static org.junit.Assert.assertThrows;
