+++ src/test/java/com/chertiavdev/bookingapp/service/impl/StripeServiceImplTest.java
@@ -1,27 +1,27 @@
 package com.chertiavdev.bookingapp.service.impl;
 
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.ACCOMMODATION_DAILY_RATE_7550;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.BOOKING_SESSION_PREFIX;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.CAN_T_RETRIEVE_SESSION_BY_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.DEFAULT_SESSION_ITEM_QUANTITY;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_SESSION_PENDING_ID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_STATUS_PAID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.PAYMENT_STATUS_UNPAID;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SAMPLE_TEST_ID_1;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_ID_QUERY_STRING;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_ID_RETRIEVAL_ERROR_TEMPLATE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_PAYMENT_CANCEL_URL;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_PAYMENT_SUCCESS_URL;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_STATUS_EXPIRED;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.SESSION_STATUS_OPEN;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.STRIPE_API_ERROR_MESSAGE;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.STRIPE_APP_BASE_URL;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.STRIPE_APP_BASE_URL_NAME;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.STRIPE_CURRENCY_NAME;
-import static com.chertiavdev.bookingapp.utils.constants.ServiceTestConstants.STRIPE_CURRENCY_USD;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACCOMMODATION_DAILY_RATE_7550;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.ACTUAL_RESULT_SHOULD_NOT_BE_NULL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.BOOKING_SESSION_PREFIX;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.CAN_T_RETRIEVE_SESSION_BY_ID;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.DEFAULT_SESSION_ITEM_QUANTITY;
 import static com.chertiavdev.bookingapp.utils.constants.TestConstants.EXCEPTION_MESSAGE_SHOULD_BE_EQUAL_TO_THE_EXPECTED_ONE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_SESSION_PENDING_ID;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_STATUS_PAID;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.PAYMENT_STATUS_UNPAID;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SAMPLE_TEST_ID_1;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_ID_QUERY_STRING;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_ID_RETRIEVAL_ERROR_TEMPLATE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_PAYMENT_CANCEL_URL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_PAYMENT_SUCCESS_URL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_STATUS_EXPIRED;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.SESSION_STATUS_OPEN;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.STRIPE_API_ERROR_MESSAGE;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.STRIPE_APP_BASE_URL;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.STRIPE_APP_BASE_URL_NAME;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.STRIPE_CURRENCY_NAME;
+import static com.chertiavdev.bookingapp.utils.constants.TestConstants.STRIPE_CURRENCY_USD;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
